# api/openapi/q2-api.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "q2 API",
    "description": "Kudos (q2) — shared self-care goals. Reference API for the initial version.",
    "version": "v1"
  },
  "paths": {
...
    "/api/goals": {
      "get": {
        "tags": [
          "Goals"
        ],
        "summary": "Lists goals, newest first.",
        "operationId": "ListGoals",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/GoalStatus"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GoalResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Goals"
        ],
        "summary": "Creates a goal.",
        "operationId": "CreateGoal",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGoalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GoalResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/goals/{id}": {
      "get": {
        "tags": [
          "Goals"
        ],
        "summary": "Returns one goal with its team and its resolved windows.",
        "operationId": "GetGoal",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GoalDetailResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Goals"
        ],
        "summary": "Deletes a stopped goal outright, for everybody on it.",
        "operationId": "DeleteGoal",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/goals/{id}/pause": {
      "post": {
        "tags": [
          "Goals"
        ],
        "summary": "Sets a goal aside for whole days, with a reason its friends read.",
        "operationId": "PauseGoal",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestPauseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GoalResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Goals"
        ],
        "summary": "Ends the running pause early.",
        "operationId": "EndGoalPause",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GoalResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/goals/{id}/pause/veto": {
      "post": {
        "tags": [
          "Goals"
        ],
        "summary": "Objects to a running pause, or takes the objection back. Anonymous.",
        "operationId": "VetoGoalPause",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GoalResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/goals/{id}/close": {
      "post": {
        "tags": [
          "Goals"
        ],
        "summary": "Stops a goal for good and moves it to the archive.",
        "operationId": "CloseGoal",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CloseGoalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GoalResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CloseGoalRequest": {
        "type": "object",
        "properties": {
          "completed": {
            "type": "boolean",
            "nullable": true
          }
        }
      },
...
      "CreateGoalRequest": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "icon": {
            "type": "string",
            "nullable": true
          },
          "schedule": {
            "oneOf": [
              {
                "enum": [
                  null
                ]
              },
              {
                "$ref": "#/components/schemas/GoalScheduleRequest"
              }
            ]
          },
          "isGroup": {
            "type": "boolean",
            "nullable": true
          },
          "reminderAt": {
            "type": "string",
            "format": "time",
            "nullable": true
          },
          "targetDate": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "participantIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "nullable": true
          }
        }
      },
...
      "GoalDetailResponse": {
        "required": [
          "goal",
          "team",
          "history"
        ],
        "type": "object",
        "properties": {
          "goal": {
            "$ref": "#/components/schemas/GoalResponse"
          },
          "team": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GoalTeamMemberResponse"
            }
          },
          "history": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GoalInstanceResponse"
            }
          }
        }
      },
      "GoalInstanceResponse": {
        "required": [
          "id",
          "startsOn",
          "dueOn",
          "dueAt",
          "requiredProofs",
          "confirmedProofs",
          "remainingProofs",
          "status",
          "pendingProofId",
          "acceptsProof"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "startsOn": {
            "type": "string",
            "format": "date"
          },
          "dueOn": {
            "type": "string",
            "format": "date"
          },
          "dueAt": {
            "type": "string",
            "format": "date-time"
          },
          "requiredProofs": {
            "type": "integer",
            "format": "int32"
          },
          "confirmedProofs": {
            "type": "integer",
            "format": "int32"
          },
          "remainingProofs": {
            "type": "integer",
            "format": "int32"
          },
          "status": {
            "$ref": "#/components/schemas/GoalInstanceStatus"
          },
          "pendingProofId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "acceptsProof": {
            "type": "boolean"
          }
        }
      },
      "GoalInstanceStatus": {
        "enum": [
          "Open",
          "Done",
          "Missed",
          "Paused"
        ]
      },
      "GoalPauseResponse": {
        "required": [
          "id",
          "reason",
          "startsOn",
          "endsOn",
          "endsAt",
          "days",
          "vetoCount",
          "vetoesRequired",
          "vetoedByMe",
          "canVeto"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "reason": {
            "type": "string"
          },
          "startsOn": {
            "type": "string",
            "format": "date"
          },
          "endsOn": {
            "type": "string",
            "format": "date"
          },
          "endsAt": {
            "type": "string",
            "format": "date-time"
          },
          "days": {
            "type": "integer",
            "format": "int32"
          },
          "vetoCount": {
            "type": "integer",
            "format": "int32"
          },
          "vetoesRequired": {
            "type": "integer",
            "format": "int32"
          },
          "vetoedByMe": {
            "type": "boolean"
          },
          "canVeto": {
            "type": "boolean"
          }
        }
      },
      "GoalResponse": {
        "required": [
          "id",
          "title",
          "description",
          "icon",
          "schedule",
          "status",
          "isGroup",
          "current",
          "streak",
          "windowsDone",
          "windowsMissed",
          "reminderAt",
          "targetDate",
          "createdAt",
          "closedAt",
          "participants",
          "isOverdue",
          "isMine",
          "risk",
          "pause",
          "remainingPauses"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "icon": {
            "type": "string"
          },
          "schedule": {
            "$ref": "#/components/schemas/GoalScheduleResponse"
          },
          "status": {
            "$ref": "#/components/schemas/GoalStatus"
          },
          "isGroup": {
            "type": "boolean"
          },
          "current": {
            "oneOf": [
              {
                "enum": [
                  null
                ]
              },
              {
                "$ref": "#/components/schemas/GoalInstanceResponse"
              }
            ]
          },
          "streak": {
            "type": "integer",
            "format": "int32"
          },
          "windowsDone": {
            "type": "integer",
            "format": "int32"
          },
          "windowsMissed": {
            "type": "integer",
            "format": "int32"
          },
          "reminderAt": {
            "type": "string",
            "format": "time",
            "nullable": true
          },
          "targetDate": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "closedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "participants": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PersonSummary"
            }
          },
          "isOverdue": {
            "type": "boolean"
          },
          "isMine": {
            "type": "boolean"
          },
          "risk": {
            "oneOf": [
              {
                "enum": [
                  null
                ]
              },
              {
                "$ref": "#/components/schemas/RiskResponse"
              }
            ]
          },
          "pause": {
            "oneOf": [
              {
                "enum": [
                  null
                ]
              },
              {
                "$ref": "#/components/schemas/GoalPauseResponse"
              }
            ]
          },
          "remainingPauses": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        }
      },
      "GoalScheduleRequest": {
        "type": "object",
        "properties": {
          "kind": {
            "oneOf": [
              {
                "enum": [
                  null
                ]
              },
              {
                "$ref": "#/components/schemas/ScheduleKind"
              }
            ]
          },
          "everyDays": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "weekdays": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Weekday"
            },
            "nullable": true
          },
          "times": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "period": {
            "oneOf": [
              {
                "enum": [
                  null
                ]
              },
              {
                "$ref": "#/components/schemas/QuotaPeriod"
              }
            ]
          }
        }
      },
      "GoalScheduleResponse": {
        "required": [
          "kind",
          "everyDays",
          "weekdays",
          "times",
          "period"
        ],
        "type": "object",
        "properties": {
          "kind": {
            "$ref": "#/components/schemas/ScheduleKind"
          },
          "everyDays": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "weekdays": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Weekday"
            }
          },
          "times": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "period": {
            "oneOf": [
              {
                "enum": [
                  null
                ]
              },
              {
                "$ref": "#/components/schemas/QuotaPeriod"
              }
            ]
          }
        }
      },
      "GoalStatus": {
        "enum": [
          "Active",
          "Completed",
          "Archived"
        ]
      },
      "GoalTeamMemberResponse": {
        "required": [
          "person",
          "streak"
        ],
        "type": "object",
        "properties": {
          "person": {
            "$ref": "#/components/schemas/PersonSummary"
          },
          "streak": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "HttpValidationProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        }
      },
...
      "PersonSummary": {
        "required": [
          "id",
          "displayName",
          "handle",
          "initials",
          "avatarColor",
          "isOnline",
          "avatarImageId"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "displayName": {
            "type": "string"
          },
          "handle": {
            "type": "string"
          },
          "initials": {
            "type": "string"
          },
          "avatarColor": {
            "type": "string"
          },
          "isOnline": {
            "type": "boolean"
          },
          "avatarImageId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          }
        }
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "QuotaPeriod": {
        "enum": [
          "Week",
          "Month",
          null
        ]
      },
...
      "RequestPauseRequest": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "nullable": true
          },
          "days": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        }
      },
      "RiskReason": {
        "enum": [
          "LastDay",
          "Tight"
        ]
      },
      "RiskResponse": {
        "required": [
          "reason",
          "missingProofs",
          "requiredProofs",
          "remainingDays"
        ],
        "type": "object",
        "properties": {
          "reason": {
            "$ref": "#/components/schemas/RiskReason"
          },
          "missingProofs": {
            "type": "integer",
            "format": "int32"
          },
          "requiredProofs": {
            "type": "integer",
            "format": "int32"
          },
          "remainingDays": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "ScheduleKind": {
        "enum": [
          "Once",
          "Interval",
          "Weekdays",
          "Times"
        ]
      },
...
      "Weekday": {
        "enum": [
          "Monday",
          "Tuesday",
          "Wednesday",
          "Thursday",
          "Friday",
          "Saturday",
          "Sunday"
        ]
      }
    }
  },
...
}